+++ tests/compile/funcs/snake_case.rs
@@ -1,6 +1,6 @@
 use compose_idents::compose_idents;
 
-compose_idents!(my_var = [snake_case(fooBar)]; {
+compose_idents!(my_var = [snake_case(fooBar)], {
     const my_var: u32 = 42;
 });