+++ tests/compile/funcs/upper.rs
@@ -1,6 +1,6 @@
 use compose_idents::compose_idents;
 
-compose_idents!(my_var = [upper(foo), _, upper(bar)]; {
+compose_idents!(my_var = [upper(foo), _, upper(bar)], {
     const my_var: u32 = 42;
 });