+++ tests/compile/nested_compose.rs
@@ -2,7 +2,7 @@ use compose_idents::compose_idents;
 
 macro_rules! outer_macro {
     ($name:tt) => {
-        compose_idents!(my_fn = [foo, _, $name]; {
+        compose_idents!(my_fn = [foo, _, $name], {
             fn my_fn() -> u32 {
                 42
             }