+++ tests/compile/num_compose.rs
@@ -1,6 +1,6 @@
 use compose_idents::compose_idents;
 
-compose_idents!(my_fn = [foo, _, 1, _, bar]; {
+compose_idents!(my_fn = [foo, _, 1, _, bar], {
     fn my_fn() -> u32 {
         42
     }