+++ tests/tests.rs
@@ -2,7 +2,7 @@
 fn compile_tests() {
     let t = trybuild::TestCases::new();
     t.pass("tests/compile/simple_compose.rs");
-    t.pass("tests/compile/trailing_semicolon.rs");
+    t.pass("tests/compile/trailing_comma.rs");
     t.pass("tests/compile/nested_compose.rs");
     t.pass("tests/compile/nested_type_token_compose.rs");
     t.pass("tests/compile/multi_compose.rs");
@@ -17,6 +17,14 @@ fn compile_tests() {
     t.pass("tests/compile/funcs/camel_case.rs");
     t.pass("tests/compile/funcs/hash.rs");
     t.pass("tests/compile/format_lit_str.rs");
+}
+
+/// Tests semicolon backwards-compatibility support.
+#[test]
+fn semicolon_tests() {
+    let t = trybuild::TestCases::new();
+    t.pass("tests/compile/semicolon/semicolon_syntax.rs");
+    t.compile_fail("tests/compile/semicolon/mixed_separators.rs");
 }
 
 #[test]